requirements.txt
# base

numpy>=1.18.5,<1.27
scipy==1.10.1
matplotlib==3.7.1
scikit-learn==1.4.1.post1
six==1.16.0
Pillow==10.2.0
tqdm==4.66.1
statsmodels==0.13.5
pydub==0.25.1
resampy==0.4.2
ffmpeg-python==0.2.0
cma==3.3.0
pandas==2.1.4
librosa==0.10.1
numba~=0.56.4
opencv-python
sortedcontainers==2.4.0
h5py==3.10.0
multiprocess>=0.70.12

# frameworks

tensorflow==2.14.0
keras==2.14.0
tensorflow-addons>=0.13.0

# using mxnet-native for reproducible test results on CI machines without Intel Architecture Processors, but mxnet is fully supported by ART
# mxnet-native==1.8.0.post0

# PyTorch
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1+cpu

# PyTorch image transformers
timm==0.9.2

catboost==1.2.3
GPy==1.13.1
lightgbm==4.3.0
xgboost==2.0.3

kornia~=0.7.1
tensorboardX==2.6.2.2
lief==0.14.1
jax[cpu]==0.4.23

# Lingvo ASR dependencies
# supported versions: (lingvo==0.6.4 with tensorflow-gpu==2.1.0)
# note: due to conflicts with other TF1/2 version supported by ART, the dependencies are not installed by default
# tensorflow-gpu==2.1.0
# lingvo==0.6.4

# other
pytest~=7.4.3
pytest-flake8~=1.1.1
flake8~=4.0.1
pytest-mock~=3.12.0
pytest-cov~=4.1.0
requests~=2.31.0

# ART
-e .

# NOTE to contributors: When changing/adding packages, please make sure that the packages are consistent with those
# present within the Dockerfile
